Extract duplicated code of the loop detectors in an abstract class.
