+++ lincheck/src/jvm/main/org/jetbrains/kotlinx/lincheck/strategy/managed/AdaptiveLoopDetector.kt
@@ -11,9 +11,8 @@
 package org.jetbrains.kotlinx.lincheck.strategy.managed
 
 import org.jetbrains.kotlinx.lincheck.util.mutableThreadMapOf
-import kotlin.math.abs
 //TODO: check for unused function parameters
-//TODO: eliminate duplicates
+
 /**
  * Loop detector that classifies loops based on observed shared-memory operations
  * and makes switching/stuck decisions accordingly.
@@ -43,10 +42,8 @@ class AdaptiveLoopDetector(
     // Threshold for abstract state revisits before declaring STUCK.
     val stuckThreshold: Int = 60,
     // Threshold for recursive method calls.
-    val recursiveCallsBound: Int = 50,
-) : LoopDetector {
-
-    private val threadStates = mutableThreadMapOf<LoopDetectorThreadState>()
+    override val recursiveCallsBound: Int = 50,
+) : AbstractLoopDetector() {
 
     // Per-loop adaptive state, keyed by (threadId, loopId, codeLocation)
     private val loopInstances = mutableThreadMapOf<MutableMap<LoopKey, LoopInstanceState>>()
@@ -60,27 +57,6 @@ class AdaptiveLoopDetector(
     // Thread that performed the last write to each location
     private val locationWriteThread = mutableMapOf<Int, Int>()
 
-    private fun currentFrame(threadId: Int): ActiveMethodCallInfo? =
-        threadStates[threadId]?.callStack?.lastOrNull()
-
-    private fun getOrCreateFrame(threadId: Int): ActiveMethodCallInfo {
-        val state = state(threadId)
-        return state.callStack.lastOrNull() ?: ActiveMethodCallInfo(methodId = -1).also {
-            state.callStack.addLast(it)
-        }
-    }
-
-    private fun findLoopInCurrentFrame(
-        threadId: Int,
-        loopId: Int,
-        codeLocation: Int,
-    ): ActiveLoopInfo? {
-        val frame = currentFrame(threadId) ?: return null
-        return frame.loops.lastOrNull {
-            it.key.loopId == loopId && it.key.codeLocation == codeLocation
-        }
-    }
-
     private fun findActiveLoop(threadId: Int): ActiveLoopInfo? {
         val state = threadStates[threadId] ?: return null
         for (frame in state.callStack.reversed()) {
@@ -90,37 +66,8 @@ class AdaptiveLoopDetector(
         return null
     }
 
-    private fun findExitLoop(
-        threadId: Int,
-        loopId: Int,
-        exitCodeLocation: Int,
-    ): Pair<ActiveMethodCallInfo, ActiveLoopInfo>? {
-        val state = threadStates[threadId] ?: return null
-        val currentMethod = getCurrentMethodId(threadId)
-
-        var best: Pair<ActiveMethodCallInfo, ActiveLoopInfo>? = null
-        var bestDistance = Int.MAX_VALUE
-
-        for (frame in state.callStack.reversed()) {
-            if (frame.methodId != currentMethod && frame.methodId != -1) continue
-
-            for (loop in frame.loops.reversed()) {
-                if (loop.key.loopId != loopId) continue
-
-                val dist = kotlin.math.abs(loop.key.codeLocation - exitCodeLocation)
-                if (dist < bestDistance) {
-                    best = Pair(frame, loop)
-                    bestDistance = dist
-                }
-                if (dist == 0)
-                    return Pair(frame, loop)
-            }
-        }
-        return best
-    }
-
     override fun resetAll() {
-        threadStates.clear()
+        super.resetAll()
         loopInstances.clear()
         globalWriteVersion = 0L
         locationWriteVersions.clear()
@@ -128,17 +75,9 @@ class AdaptiveLoopDetector(
     }
 
     override fun resetThread(threadId: Int) {
-        threadStates.remove(threadId)
+        super.resetThread(threadId)
         loopInstances.remove(threadId)
     }
-
-    private fun state(threadId: Int): LoopDetectorThreadState =
-        threadStates.getOrPut(threadId) {
-            LoopDetectorThreadState(threadId)
-        }
-
-    override fun getCurrentMethodId(threadId: Int): Int =
-        currentFrame(threadId)?.methodId ?: -1
 
     private fun instances(threadId: Int): MutableMap<LoopKey, LoopInstanceState> =
         loopInstances.getOrPut(threadId) { mutableMapOf() }
@@ -153,39 +92,14 @@ class AdaptiveLoopDetector(
         return instances(threadId)[loop.key]
     }
 
-    override fun getCurrentLoopIteration(threadId: Int, loopId: Int, codeLocation: Int): Int {
-        val state = state(threadId)
-        val frame = state.callStack.lastOrNull() ?: return 1
-        val loop = frame.loops.lastOrNull { it.key.loopId == loopId && it.key.codeLocation == codeLocation } ?: return 1
-        return loop.iterationCount
-    }
-
     // --- LOOP LEVEL ---
-    override fun beforeLoopEnter(threadId: Int, codeLocation: Int, loopId: Int) {
-        val frame = getOrCreateFrame(threadId)
-        val existing = frame.loops.lastOrNull {
-            it.key.loopId == loopId && it.key.codeLocation == codeLocation
-        }
-        if (existing == null) {
-            frame.loops.addLast(ActiveLoopInfo(LoopKey(loopId, codeLocation)))
-        }
-    }
-
     override fun onLoopIteration(
         threadId: Int,
         codeLocation: Int,
         loopId: Int
     ): Pair<Boolean, LoopDetector.Decision> {
-        var loop = findLoopInCurrentFrame(threadId, loopId, codeLocation)
-
-        val started = if (loop == null) {
-            beforeLoopEnter(threadId, codeLocation, loopId)
-            loop = findLoopInCurrentFrame(threadId, loopId, codeLocation)
-            true
-        } else {
-            false
-        }
-        loop!!.iterationCount += 1
+        val (loop, started) = startLoopIfNeeded(threadId, codeLocation, loopId)
+        loop.iterationCount += 1
 
         val inst = getOrCreateInstance(threadId, loopId, codeLocation)
         val decision = processIteration(inst)
@@ -212,19 +126,11 @@ class AdaptiveLoopDetector(
         loopId: Int,
         isReachableFromOutsideLoop: Boolean
     ): Int? {
-        val match = findExitLoop(threadId, loopId, codeLocation)
-
-        if (!isReachableFromOutsideLoop || match != null) {
-            val enterCodeLocation = match?.second?.key?.codeLocation ?: codeLocation
-            match?.let { (frame, loop) ->
-                loop.iterationCount = 0
-                frame.loops.remove(loop)
-            }
-
+        val enterCodeLocation = super.afterLoopExit(threadId, codeLocation, loopId, isReachableFromOutsideLoop)
+        if (enterCodeLocation != null) {
             instances(threadId).remove(LoopKey(loopId, enterCodeLocation))
-            return enterCodeLocation
         }
-        return null
+        return enterCodeLocation
     }
 
     // --- Decision logic ---
@@ -398,64 +304,6 @@ class AdaptiveLoopDetector(
         return h
     }
 
-    // --- METHOD LEVEL ---
-    override fun onMethodEnter(
-        threadId: Int,
-        codeLocation: Int,
-        methodId: Int,
-        receiver: Any?,
-        params: Array<Any?>
-    ): LoopDetector.Decision {
-        val st = state(threadId)
-        val stack = st.callStack
-        val top = stack.lastOrNull()
-
-        val counters = st.methodCallCounters
-        val count = (counters[methodId] ?: 0) + 1
-        counters[methodId] = count
-        if (count > recursiveCallsBound) {
-            return LoopDetector.Decision.STUCK
-        }
-
-        if (top != null && top.methodId == methodId) {
-            top.depth += 1
-            if (top.depth > recursiveCallsBound) {
-                return LoopDetector.Decision.STUCK
-            }
-        } else {
-            stack.addLast(ActiveMethodCallInfo(methodId = methodId))
-        }
-
-        return LoopDetector.Decision.IDLE
-    }
-
-    override fun onMethodExit(
-        threadId: Int,
-        methodId: Int,
-        receiver: Any?,
-        params: Array<Any?>,
-        result: Any?
-    ) {
-        val st = state(threadId)
-        val stack = st.callStack
-        val top = stack.lastOrNull() ?: return
-
-        val counters = st.methodCallCounters
-        val count = (counters[methodId] ?: 1) - 1
-        if (count <= 0) counters.remove(methodId) else counters[methodId] = count
-
-        if (top.methodId == methodId) {
-            if (top.depth > 1) top.depth -= 1 else stack.removeLast()
-            return
-        }
-
-        val idx = stack.indexOfLast { it.methodId == methodId }
-        if (idx == -1) return
-        while (stack.size - 1 > idx) stack.removeLast()
-        val frame = stack.lastOrNull() ?: return
-        if (frame.depth > 1) frame.depth -= 1 else stack.removeLast()
-    }
-
     // --- IRREDUCIBLE LOOPS ---
     override fun onIrreducibleLoopIteration(threadId: Int, codeLocation: Int, loopId: Int): LoopDetector.Decision {
         val inst = getOrCreateInstance(threadId, loopId, codeLocation)
@@ -463,7 +311,7 @@ class AdaptiveLoopDetector(
         val state = threadStates[threadId]
         val frame = state?.callStack?.lastOrNull()
         val loop = frame?.loops?.lastOrNull { it.key.loopId == loopId && it.key.codeLocation == codeLocation }
-        loop?.iterationCount = (loop.iterationCount ?: 0) + 1
+        loop?.iterationCount = (loop?.iterationCount ?: 0) + 1
 
         return processIteration(inst)
     }
