+++ lincheck/src/jvm/main/org/jetbrains/kotlinx/lincheck/strategy/managed/ManagedStrategy.kt
@@ -81,7 +81,7 @@ internal abstract class ManagedStrategy(
 //        recursiveCallsBound = settings.recursionBound,
 //    )
 
-    internal val loopDetector: LoopDetector = AdaptiveLoopDetector()
+    internal val loopDetector: LoopDetector = AdaptiveLoopDetector(recursiveCallsBound = settings.recursionBound)
     // Current execution part, if defined by the runner, `PARALLEL` otherwise
     protected val currentExecutionPart: ExecutionPart
         get() = (runner as? ExecutionScenarioRunner)?.currentExecutionPart ?: PARALLEL
